+++ app/src/main/java/com/unaerp/restaurantmenu/Domain/ResponseShoppingCartItem.kt
@@ -1,12 +1,12 @@
 package com.unaerp.restaurantmenu.Domain
 
 class ResponseShoppingCartItem(
-    var id: String, var quantity: Int
+    var id: String, var quantity: Long
 ) {
     companion object {
         fun fromMap(map: Map<String, Any>): ResponseShoppingCartItem {
             return ResponseShoppingCartItem(
-                (map["id"] ?: "") as String, (map["quantity"] ?: 0) as Int
+                (map["id"] ?: "") as String, (map["quantity"] ?: 0) as Long
             )
         }
     }
