+++ app/src/main/java/com/unaerp/restaurantmenu/core/data_source/RemoteUserDataSource.kt
@@ -12,5 +12,5 @@ interface RemoteUserDataSource {
 
     suspend fun finishPurchase(id: String): OnResult<Unit>
 
-    suspend fun setProductCart(idUser: String, idProduct: String, quantity: Int): OnResult<Unit>
+    suspend fun setProductCart(idUser: String, idProduct: String, quantity: Long): OnResult<Unit>
 }
